updated primary key (id) type from integer to long throughout the application. additionally added delete application route

+++ src/main/java/com/memento/practice/api/exceptions/ApplicationNotFoundException.java
@@ -1,7 +1,7 @@
 package com.memento.practice.api.exceptions;
 
 public class ApplicationNotFoundException extends RuntimeException {
-    public ApplicationNotFoundException(Integer id) {
+    public ApplicationNotFoundException(Long  id) {
         super("Application not found with id " + id);
     }
 }
