+++ src/main/java/com/memento/practice/api/repositories/JobApplicationRepository.java
@@ -6,5 +6,5 @@ import com.memento.practice.api.models.JobApplication;
 
 //jpa repository  uses generics, you pass your created Schema (class)
 //and then the type  of the primary key
-public interface JobApplicationRepository extends JpaRepository<JobApplication, Integer>{
+public interface JobApplicationRepository extends JpaRepository<JobApplication, Long>{
 }
